feat: resolução do desafio

+++ src/controller/UserController.js
@@ -5,6 +5,7 @@ const Sequelize = require("sequelize");
 module.exports = {
   async index(req, res) {
     try {
+      console.log(req.user);
       if (req.user.nivel === 999 && req.user.acesso === 1) {
         const users = await User.findAll();
         return res.json(users);
@@ -17,9 +18,26 @@ module.exports = {
         .json({ error: "Não foi possivel mostrar os usuarios" });
     }
   },
+  async show(req, res) {
+    try {
+      const { id } = req.params;
+      if (req.user.nivel === 999 || req.user.id === id) {
+        const user = await User.findOne({
+          where: { id },
+        });
+        return res.status(200).json(user);
+      }
+      return res.status(401).json({ error: "Você não tem autorização" });
+    } catch (err) {
+      return res
+        .status(400)
+        .json({ error: "Não foi possivel criar o usuario" });
+    }
+  },
   async store(req, res) {
     try {
       const data = req.body;
+
       const userExist = await User.findOne({
         where: Sequelize.or({ email: data.email }, { cpf: data.cpf }),
       });
@@ -40,24 +58,16 @@ module.exports = {
     try {
       const { id } = req.params;
       const data = req.body;
-      console.log(req.user);
-      const userExist = await User.findOne({
-        where: Sequelize.or({ email: data.email }, { cpf: data.cpf }),
-      });
-      if (userExist) {
-        if (userExist.email === data.email) {
-          return res.status(400).json({ error: "E-mail já cadastrado" });
-        }
-        return res.status(400).json({ error: "CPF já cadastrado" });
-      }
       if (data.senha) {
         data.senha = await bcrypt.hash(data.senha, 8);
       }
-      const user = await User.update(data, {
-        where: { id },
-      });
-
-      return res.json({ message: "Atualizado com sucesso!" });
+      if (req.user.nivel === 999 || req.user.id === id) {
+        await User.update(data, {
+          where: { id },
+        });
+        return res.json({ message: "Atualizado com sucesso!" });
+      }
+      return res.status(401).json({ error: "Você não tem autorização" });
     } catch (err) {
       console.log(err);
       return res.json({ error: "Não foi possivel atualizar o usuario" });
@@ -66,10 +76,13 @@ module.exports = {
   async destroy(req, res) {
     try {
       const { id } = req.params;
-      const user = await User.destroy({
-        where: { id },
-      });
-      return res.json({ message: "Deletado com sucesso!" });
+      if (req.user.nivel === 999 || req.user.id === id) {
+        await User.destroy({
+          where: { id },
+        });
+        return res.json({ message: "Deletado com sucesso!" });
+      }
+      return res.status(401).json({ error: "Você não tem autorização" });
     } catch (err) {
       return res.json({ error: "Não foi possivel deletar o usuario" });
     }
